Remove container pattern and use new redux-connect hooks
