+++ generators/base/templates/App/Screens/index.js
@@ -1,6 +1,6 @@
 // @flow
-import Main from '<%= name %>/App/Screens/Main';
-import Styleguide from '<%= name %>/App/Screens/Styleguide/Container';
+import Main from "<%= name %>/App/Screens/Main";
+import Styleguide from "<%= name %>/App/Screens/Styleguide";
 
 export default {
   Main,
